Refactor output directory functionality
First, a single term 'output directory' is introduced in both functions,
variables and description. Arguably, this term is clearer for the user
than 'working directory', so why not use it in the code too.
Second, the field was actually called 'input_directory' in the dialog
file for some reason. It was also renamed there.

An important change is that now the reference to the output directory is
stored in the main class. At the start, it is initialized from cache
(QSettings), and if empty, a relevant dialog pops up. Previously, its
value was accesses alternately from the form itself and from QSettings.