Use plugin toolbar instead of own
For now, the plugin is meant to have a single button, which makes creating
a separate toolbar redundant. Another minor problem is that a custom toolbar
is always added at a new line, and not at the end of the last one. This
slightly decreases the space left for the canvas and other panels.