Restrict AOI combo to layers with a single feature
Previously, all polygon layers were added to the combo. However, the behavior in the case
of a layer with multiple features or a layer with no features was undefined.
It would eventually result in server error and had to be handled. But it is currently
arguably more intuitive to not even let the user use those layers to prevent errors instead of handling them.