Implement processing refresh using QgsTask
QgsTask is a wrapper around Thread.
Unlike Python Thread itself, it has convenient hooks and signals that come in handy here.
The blocking work, namely HTTP request, is done in the task.
The GUI work (filling out the table) is done in the main thread in a separate function that is triggered
by the task's progress change.

However, there are notable drawbacks. QgsTask inevitably displays progress bars and upon-completion pop-ups
to the user, while in this case the task should remain invisible for the user.

QThread may be used instead in the future.