Reset raster combo to Mapbox if no TIF selected
To select a local GeoTIFF, the user select the option 'Open new .tif' in the relevant combo.
If they select an image, the combo's index moves to the new image.
However, if they just quit the file selection dialog, the combo remains at 'Open new .tif', but it's functional anymore.
If they want to select a new TIF again, they'd have to click the combo, click another option, then click on the TIF one again.

Now, the combo's automatically reset to index 0 (Mapbox) in that case.