Do not load empty processing results
PyQGIS raises an error when trying to create a vector layer with no features. At least via OGR.
Since the user will hardly need the output if it's empty, the results are not loaded at all.
Instead, a push message is shown notifying the user that the processing produced no results.