improvement: use QGIS auth for custom provider
QGIS authentication system is based on an SQLite DB stored in the QGIS user profile directory.
User creates an authentication config, i.e. an entity that includes the credentials along with the authentication method.
It is essentially an entry in the database. The login & password themselves are stored encrypted.
To retireve and/or decrypt the credentials, a master password is required. This is a password the user creates beforehand.

In the UI, the auth system is represented by a widget that allows selecting an existing auth config, or creating a new one.

It's worth noting that there's another auth widget used across QGIS that actually allows circumventing this system
by simply saving the credentials unencrypted in the project file, or not saving them at all.
But this widget appears to be unavailable for plugin creation (not part of the relevant .so library).
So, unlike various other core parts, will enforce using QGIS Auth DB.
But I believe that its benefits outweight the possible overhead for the user.

For more details, see:
https://docs.qgis.org/3.16/en/docs/pyqgis_developer_cookbook/authentication.html