fix(timeouts): remove TIF upload timeout
TIFs can greatly vary in size, and so can their upload time,
not to mention that it also depends on the user's network connection.

Besides, the upload happens in a separate thread, so it won't block the UI.

There may be a need for a timeout, but that should be some large value that has  yet to be determined.