feat(output-dir): open dir selection dlg if no dir
Previously, if the called required an output dir, but it wasn't set,
an alert was shown to the user and the method quit.
To better prompt the user to select an output directory, the selection
dialog now opens automatically in such cases.
If the user quit the dialog without selecting a dir, the same alert is shown.