chore!: set min QGIS version to 3.16
Min version should logically correspond to one of the long-term releases (LTR).
At the time of this commit, LTR is 3.16, and the one before was 3.10.
Since plugin was developed using 3.16+, some of it's code is incompatible with earlier versions.
Although we suppose that many users may still use 3.10, we believe it is reasonable
to force them to upgrade to 3.16 since it's LTR, and it's been around for 7 months.