feat: add support for QGIS 3.10+
The only incompatibility with QGIS 3.10, 3.12 & 3.14 seemed to be the
QgsVectorFileWriter's method writeAsVectorFormatV2 that was only introduced in 3.10.3.
Before that, only writeAsVectorFormat was present.
Starting with 3.20, writeAsVectorFormatV3 is added and the rest are set to be deprecated.
Although no visible difference in the way work is observed, this commit adds support for
all them by checking the QGIS version.