feat(maxar): get FID from table selection
Previously, in order to process a specific image, user had to select a row in the
metadata table, click 'Get Feature ID', and then 'Get URL' . And there existed a
field for that FID. Since in fact the user won't need to see the FID itself,
this intermediate step can be skipped by just adding the FID to the URL when 'Get URL'
is clicked, provided a row in the table is selected.
The table selection model is set to single-row, so there will always be a single
FID selected.