fix: do img&aoi intersection only if Custom Provider
There used to be no check for the raster combo item if the Maxar metadata table was
enabled and had selection. It was assumed that under these circumstances, the user
uses Maxar (Custom provider) for processing.
In fact, a user may play around with the metadata table, leave it on, but
switch to another imagery source, which will cause an error if the user's AOI
doesn't intersects with the selected Maxar image.