feat(maxar): check ConnectID validity
Display an alert and abort if ConnectID field is empty or has a value
that doesn't make a valid UUID.

An invalid or missing ConnectID somehow doesn't result in an HTTP Error.
Instead, the request succeeds but the resulting metadata layer is empty.
This commit makes the problem more explicit to the user.