feat(maxar): let edit connect ID via dedicated dialog
Maxar Connect IDs are now selected from a read-only dropdown list with the three services fixed:
 - SecureWatch
 - Vivid
 - Basemaps

There's no option to add or remove to or from it, but only to reassign the ids via the
"Edit Connect ID" button next to it. It opens up a minimalistic dialog with a
line edit to enter the new id and the standard button to close the dialog.
The input is validated. If the text isn't a vaid UUID, the line edit gets outlined with red.