Add config for custom providers (#8)
* Add provider config

Web imagery providers are now managed via a dropdown list. The list can be edited
via the standard 'add', 'remove' and 'edit' buttons next to the list widget.
The add button opens up a dialog with a form where the user can specify the provider
name as it will appear in the list, URL, and provider type.
The update button opens the same dialog, but the form inputs are used to update
the provider currently selected in the list. However, if the provider is one of the
built-in Maxar products, then another dialog pops up, this time only offering
a field to type in a new Connect ID.

Internally, the providers are written to custom_providers.json in the plugin directory.

Co-authored-by: Grigory Nedaev <g.nedaev@geoalert.io>