fix: use plugin name, not file name as tr context in mapflow.py
For some reason, pylupdate5 changes it back to 'Mapflow' which must
be captialized module name w/out the suffix, because, which coincides with
the plugin name in config. And this breaks the trs added after context name change.
Since it's no big deal, just changed it back to how it used to be.