Missing preview (#20)
Preview URLs for Sentinel are mostly missing for 2019 and earlier (None), along with thumbnail URLs. And since they don't provide the original Sentinel image ID as a separate attribute, I used to extract it from the preview URL. So when it was missing, not only could not the user preview the image but neither could they start a processing since the ID was missing too.

Turns out, Mapflow doesn't actually need the full Sentinel ID, and the cell id and date are enough. Fortunately, that information can be extracted from the product_name attribute, and that seems to always be present. So now Sentinel IDs in the metadata table are of the format '/37/U/DB/2022/2/8/0/' which is passed in that to the platform. But Mapflow can still accept the full image IDs too, because the user may already have such an ID and just paste it into the Image ID field.