Change Maxar meta request from GET to POST
Initially, we requested meta by BBOX, but that meant that some of the fetched images might intersects the bbox but not the AOI itself, and the plugin had to additionally discard those image by doing spatial checks itself. To move that computationally intensive procedure to Maxar, I started using a CQL filter in the URL where I put the AOI itself, and not its bbox. But the consequence was that a geometry with many vertices would produce an overly long URL and result in HTTP 413. To tackle this but keep sending the exact AOI as a filter, I changed the request to POST and CQL filter to an OGC XML filter.