fix: remove pitfalls in error handling
Previous commit improved how the exception handled by the backend are presented
(those returned as JSON) but I forgot to consider the unhandled exception for which the backend
simply hands over the source error message as plain text.

Additionally, in the latter case if the error message exceeded ~5000 characters,
an email report wouldn't work out throwing 400 due to the body size.
I now set a 5000 char limit on the size of the error message.